refactor(genui): break ui_models/schema_validation import cycle
schema_validation.dart imported ui_models.dart only to throw
A2uiValidationException, while ui_models.dart imported schema_validation.dart
for SurfaceDefinition.validate. That cycle failed layerlens --fail-on-cycles.

A2uiValidationException is a cross-cutting error type thrown across the model,
engine, and transport layers, and ui_models.dart never used the one it defined.
Move it to a primitives leaf so schema_validation no longer depends on the model
layer and the graph becomes an acyclic DAG.

primitives.dart re-exports it, so package:genui consumers are unchanged.
