+++ packages/genui/lib/src/model/ui_models.dart
@@ -236,34 +236,6 @@ final class Component {
       Object.hash(id, type, const DeepCollectionEquality().hash(properties));
 }
 
-/// Exception thrown when an A2UI message fails parsing or validation.
-class A2uiValidationException implements Exception {
-  /// Creates an [A2uiValidationException].
-  A2uiValidationException(
-    this.message, {
-    this.surfaceId,
-    this.path,
-    this.json,
-    this.cause,
-  });
-
-  final String message;
-  final String? surfaceId;
-  final String? path;
-  final Object? json;
-  final Object? cause;
-
-  @override
-  String toString() {
-    final buffer = StringBuffer('A2uiValidationException: $message');
-    if (surfaceId != null) buffer.write(' (surface: $surfaceId)');
-    if (path != null) buffer.write(' (path: $path)');
-    if (cause != null) buffer.write('\nCause: $cause');
-    if (json != null) buffer.write('\nJSON: $json');
-    return buffer.toString();
-  }
-}
-
 /// Surface lifecycle events emitted by `SurfaceController.surfaceUpdates`.
 sealed class SurfaceUpdate {
   const SurfaceUpdate(this.surfaceId);
